FIX: Hard cast to double for Statistics

+++ kernel/solver/iterative.hpp
@@ -379,7 +379,7 @@ namespace FEAST
         if(calc_def)
           this->_def_cur = this->_calc_def_norm(vec_def, vec_sol);
 
-        Statistics::add_solver_defect(this->_branch, this->_def_cur);
+        Statistics::add_solver_defect(this->_branch, double(this->_def_cur));
 
         // plot?
         if(this->_plot)
